Revise EnforcedProtocol to functional Akka Typed
Simplified the interface to protocol enforcement by making a single
function, checkTransition, do all the work. This implementation style
uses behavior functions in an object instead of an actor object. It
also provides a main/guardian actor and a simpler set of message. The
test case was improved but does not currently show the implementation
as working. This will be worked on next.